% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba.R
\docType{data}
\name{nba}
\alias{nba}
\title{NBA dataset}
\format{
A data frame with 24691 rows and 52 variables.
}
\source{
\url{}
}
\usage{
nba
}
\description{
A dataset containing NBA statistics
}
\keyword{datasets}
