% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year_cor.R
\name{year_cor}
\alias{year_cor}
\title{NBA correlation variables}
\usage{
year_cor(x)
}
\arguments{
\item{x}{The year of which you want to know statistics.}
}
\description{
This function plot the numeric continuos variables of the NBA dataset for the selected year
}
\examples{
year_cor(1996)
}
\keyword{NBA}
\keyword{correlation}
\keyword{player}
