% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year_oldest.R
\name{year_oldest}
\alias{year_oldest}
\title{Retrieve the oldest NBA player per year}
\usage{
year_oldest(x)
}
\arguments{
\item{x}{The year of which you want to know statistics.}
}
\description{
This function retrieves the oldest active player in the NBA for the year indicated
}
\examples{
year_oldest(2002)
}
\keyword{NBA}
\keyword{old}
\keyword{player}
