% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year_points.R
\name{year_points}
\alias{year_points}
\title{Function to retrieve the top 10 NBA scorers per year}
\usage{
year_points(x)
}
\arguments{
\item{x}{The year of which you want to know statistics.}
}
\description{
This function extracts the top 10 scorers in the NBA for the year indicated
}
\examples{
year_points(2015)
}
\keyword{NBA,}
\keyword{points,}
\keyword{scorers}
